refactor Home and Movie details screens
