+++ src/constants.ts
@@ -1,4 +1,9 @@
+import { Dimensions } from "react-native";
+
 export const SEARCH_PLACEHOLDER = "Search for your Favorite movies";
 export const NO_RESULTS_TEXT = "No Result";
 export const NO_MORE_RESULTS_TEXT = "No More results";
 export const ERROR_TOAST_MESSAGE = "An error has occurred. Please try later.";
+
+export const { width, height } = Dimensions.get("window");
+export const PAGE_SIZE = 10;
